shop/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-admin-thumbnails==0.2.5
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
